requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Admin
openai
pyyaml
jinja2
psycopg2-binary
